Ref: Created user_tx_token subfolder and moved files inside. Various reordering and movements of files accross subfolders
