+++ app/src/components/common/index.ts
@@ -21,7 +21,7 @@ export { Paragraph } from './text/paragraph'
 export { RadioInput } from './form/radio_input'
 export { SectionTitle } from './text/section_title'
 export { Select } from './form/select'
-export { Spinner } from './spinner'
+export { Spinner } from '../loading/spinner'
 export { SubsectionTitle } from './text/subsection_title'
 export { SubsectionTitleAction } from './text/subsection_title_action'
 export { SubsectionTitleWrapper } from '../market/sections/common_sections/top_card_details/subsection_title_wrapper'
