--- app/src/components/common/spinner/index.tsx
@@ -1,35 +1,0 @@
-import React, { HTMLAttributes } from 'react'
-import styled, { keyframes } from 'styled-components'
-
-import { IconSpinner, IconSpinnerBig } from '../icons'
-
-const rotate = keyframes`
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-`
-
-const RotatingSpinner = styled.div`
-  animation: ${rotate} 2s linear infinite;
-  flex-grow: 0;
-  flex-shrink: 0;
-`
-
-interface Props extends HTMLAttributes<HTMLDivElement> {
-  color?: string
-  size?: string
-  big?: boolean | undefined
-}
-
-export const Spinner: React.FC<Props> = (props: Props) => {
-  const { big, color = '#fff', size, ...restProps } = props
-
-  return (
-    <RotatingSpinner color={color} {...restProps}>
-      {big ? <IconSpinnerBig /> : <IconSpinner spinnerSize={size ? size : '40'} />}
-    </RotatingSpinner>
-  )
-}
