+++ app/src/components/market/sections/common_sections/top_card_details/market_top_details_closed/index.tsx
@@ -12,7 +12,7 @@ import { getContractAddress, getNativeAsset, getWrapToken } from '../../../../..
 import { getMarketRelatedQuestionFilter, onChangeMarketCurrency } from '../../../../../../util/tools'
 import { MarketMakerData, MarketState, Token } from '../../../../../../util/types'
 import { SubsectionTitleWrapper } from '../../../../../common'
-import { CurrencySelector } from '../../../../common/currency_selector'
+import { CurrencySelector } from '../../user_transactions_tokens/currency_selector'
 import { AdditionalMarketData } from '../additional_market_data'
 import { MarketData } from '../market_data'
 import { ProgressBar } from '../progress_bar'
