+++ app/src/components/market/sections/common_sections/user_transactions_tokens/transaction/transaction_details_line/index.tsx
