+++ app/src/components/market/sections/market_create/steps/items/ask_question_step.tsx
@@ -13,11 +13,11 @@ import { DateField, FormRow, FormStateButton } from '../../../../../common'
 import { BigNumberInputReturn } from '../../../../../common/form/big_number_input'
 import { CommonDisabledCSS } from '../../../../../common/form/common_styled'
 import { QuestionInput } from '../../../../../common/form/question_input'
-import { Categories } from '../../../../common/categories'
 import { MarginsButton } from '../../../../common/common_styled'
 import { CreateCard } from '../../../../common/create_card'
 import { Arbitrators } from '../../../common_sections/arbitrators/arbitrators'
 import { WarningMessage } from '../../../common_sections/messages/warning_message'
+import { Categories } from '../../../common_sections/search_filter_categories/categories'
 import { Outcome, Outcomes } from '../outcomes'
 
 import { CreateScalarMarket } from './create_scalar_market'
