+++ app/src/components/market/sections/market_sell/scalar_market_sell.tsx
@@ -46,9 +46,12 @@ import { SwitchTransactionToken } from '../../common/switch_transaction_token'
 import { GridTransactionDetails } from '../common_sections/bottom_card_details/grid_transaction_details'
 import { MarketScale } from '../common_sections/bottom_card_details/scalar/market_scale'
 import { WarningMessage } from '../common_sections/messages/warning_message'
-import { TransactionDetailsCard } from '../common_sections/transaction/transaction_details_card'
-import { TransactionDetailsLine } from '../common_sections/transaction/transaction_details_line'
-import { TransactionDetailsRow, ValueStates } from '../common_sections/transaction/transaction_details_row'
+import { TransactionDetailsCard } from '../common_sections/user_transactions_tokens/transaction/transaction_details_card'
+import { TransactionDetailsLine } from '../common_sections/user_transactions_tokens/transaction/transaction_details_line'
+import {
+  TransactionDetailsRow,
+  ValueStates,
+} from '../common_sections/user_transactions_tokens/transaction/transaction_details_row'
 
 const StyledButtonContainer = styled(ButtonContainer)`
   justify-content: space-between;
